+++ src/components/employees/employees.js
@@ -2,6 +2,7 @@ import React from 'react'
 import axios from '../../config/axios'
 import { Link } from 'react-router-dom'
 import EmployeeForm from './employees-form'
+import Swal from 'sweetalert2'
 
 class Employees extends React.Component{
     constructor(){
@@ -29,20 +30,38 @@ class Employees extends React.Component{
     handleRemove=(id)=>{
         //const data=e.target.value
         console.log(id);
-        axios.delete(`/employees/${id}`,{
-            headers:{
-                'x-auth':localStorage.getItem('authToken')
-            }
-        })
-        .then(res=>{
-            console.log(res.data)
-            this.setState((prevState)=>({
-                employees:prevState.employees.filter(emp=>{
-                    return emp._id != res.data._id 
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.value) {
+                axios.delete(`/employees/${id}`,{
+                    headers:{
+                        'x-auth':localStorage.getItem('authToken')
+                    }
                 })
-            }))
+                .then(res=>{
+                    console.log(res.data)
+                    this.setState((prevState)=>({
+                        employees:prevState.employees.filter(emp=>{
+                            return emp._id != res.data._id 
+                        })
+                    }))
+                
+                })
+                Swal.fire(
+                'Deleted!',
+                'Your file has been deleted.',
+                'success'
+              )
+            }
+          })
         
-        })
     }
     render(){
         console.log(this.state.employees);
@@ -50,7 +69,7 @@ class Employees extends React.Component{
         return(
             <div>
                 <h1>Employees List - {this.state.employees.length}</h1>
-                <ul>
+                {/* <ul>
                     {
                         this.state.employees.map(emp=>{
                             return <li key={emp._id}>
@@ -62,7 +81,33 @@ class Employees extends React.Component{
                                     </li>
                         })
                     }
-                </ul>
+                </ul> */}
+                <table className="table">
+                    <thead>
+                        <tr>
+                        <th scope="col">#</th>
+                        <th scope="col">Name</th>
+                        <th scope="col">Email</th>
+                        <th scope="col">Department</th>
+                        <th scope="col">Button</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        {this.state.employees.map((emp,i)=>{
+                            return(
+                            <tr key={emp._id}>
+                                <td>{i+1}</td>
+                                <td><Link to={`/employees/${emp._id}`}>  {emp.name}</Link></td>
+                                <td>{emp.email}</td>
+                                <td>{emp.department.name}</td>
+                                <td>
+                                <button className='btn btn-sm btn-danger' onClick={()=>{this.handleRemove(emp._id)}}>Remove</button>
+                                </td>
+                            </tr>
+                            )
+                        })}
+                    </tbody>
+                </table>
                 <Link to='/employees/new'>Add Employees</Link>
             </div>
         )
